# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  language: string;
  /** @nullable */
  pythonVersion?: string | null;
  createdAt: string;
  updatedAt: string;
  nodeCount: number;
  edgeCount: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description: string;
  language?: string;
  pythonVersion?: string;
}

export interface GraphNode {
  id: string;
  projectId: number;
  /** folder | file | class | function */
  nodeType: string;
  label: string;
  filePath: string;
  /** @nullable */
  parentNodeId?: string | null;
  positionX: number;
  positionY: number;
  width: number;
  height: number;
  inputTypes?: string[];
  outputTypes?: string[];
  /** @nullable */
  lineNumber?: number | null;
  /** @nullable */
  complexity?: number | null;
  isCollapsed?: boolean;
}

export interface GraphEdge {
  id: string;
  projectId: number;
  sourceNodeId: string;
  targetNodeId: string;
  sourceHandle: string;
  targetHandle: string;
  /** dataflow | import | inheritance */
  edgeType: string;
  /** @nullable */
  dataType?: string | null;
}

export interface ProjectGraph {
  projectId: number;
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export interface NodeMetadata {
  type: string;
  definedIn: string;
  inputs: string;
  outputs: string;
  /** @nullable */
  usedBy?: string | null;
  complexity: string;
  lastModified: string;
}

export interface NodeDetail {
  node: GraphNode;
  code: string;
  metadata: NodeMetadata;
}

export interface NodePositionUpdate {
  positionX?: number;
  positionY?: number;
  isCollapsed?: boolean;
}

export interface EdgeInput {
  sourceNodeId: string;
  targetNodeId: string;
  sourceHandle: string;
  targetHandle: string;
  edgeType?: string;
  dataType?: string;
}

export interface CodeContent {
  /** @nullable */
  nodeId: string | null;
  filePath: string;
  code: string;
  language: string;
}

export interface ProjectStats {
  projectId: number;
  totalNodes: number;
  totalEdges: number;
  folderCount: number;
  fileCount: number;
  classCount: number;
  functionCount: number;
  avgComplexity: number;
  highComplexityNodes: number;
}

export interface ActivityEvent {
  id: number;
  projectId: number;
  /** node_selected | edge_created | edge_deleted | code_edited | node_moved */
  eventType: string;
  description: string;
  /** @nullable */
  nodeId: string | null;
  timestamp: string;
}