# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  id: number;
  projectId: number;
  /** node_selected | edge_created | edge_deleted | code_edited | node_moved */
  eventType: string;
  description: string;
  /** @nullable */
  nodeId: string | null;
  timestamp: string;
}