# lib/api-zod/src/generated/types/edgeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface EdgeInput {
  sourceNodeId: string;
  targetNodeId: string;
  sourceHandle: string;
  targetHandle: string;
  edgeType?: string;
  dataType?: string;
}