# lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface GraphEdge {
  id: string;
  projectId: number;
  sourceNodeId: string;
  targetNodeId: string;
  sourceHandle: string;
  targetHandle: string;
  /** dataflow | import | inheritance */
  edgeType: string;
  /** @nullable */
  dataType?: string | null;
}