# lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface GraphNode {
  id: string;
  projectId: number;
  /** folder | file | class | function */
  nodeType: string;
  label: string;
  filePath: string;
  /** @nullable */
  parentNodeId?: string | null;
  positionX: number;
  positionY: number;
  width: number;
  height: number;
  inputTypes?: string[];
  outputTypes?: string[];
  /** @nullable */
  lineNumber?: number | null;
  /** @nullable */
  complexity?: number | null;
  isCollapsed?: boolean;
}