# lib/api-zod/src/generated/types/nodeDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphNode } from './graphNode';
import type { NodeMetadata } from './nodeMetadata';

export interface NodeDetail {
  node: GraphNode;
  code: string;
  metadata: NodeMetadata;
}