# lib/api-zod/src/generated/types/nodePositionUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface NodePositionUpdate {
  positionX?: number;
  positionY?: number;
  isCollapsed?: boolean;
}