# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  description: string;
  language: string;
  /** @nullable */
  pythonVersion?: string | null;
  createdAt: string;
  updatedAt: string;
  nodeCount: number;
  edgeCount: number;
}