# lib/api-zod/src/generated/types/projectGraph.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdge } from './graphEdge';
import type { GraphNode } from './graphNode';

export interface ProjectGraph {
  projectId: number;
  nodes: GraphNode[];
  edges: GraphEdge[];
}