# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Servioen Visual Data Engineering IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  projectId: number;
  totalNodes: number;
  totalEdges: number;
  folderCount: number;
  fileCount: number;
  classCount: number;
  functionCount: number;
  avgComplexity: number;
  highComplexityNodes: number;
}